+++ src/minifycss/index.d.ts
@@ -1,0 +1,1 @@
+export function minifycss(css:string):string